requirements.txt
Flask>=3.0,<4.0
flask-cors>=4.0,<5.0
reportlab>=4.0,<5.0
Pillow>=10.0,<12.0
gunicorn>=22.0,<23.0
pypdf>=4.2,<6.0
python-docx>=1.1,<2.0
scikit-learn>=1.4,<2.0
